Working on integrating v6/v7 entity capabilities

+++ dis_rs/src/common/errors.rs
@@ -2,6 +2,6 @@
 pub enum DisError {
     // UnsupportedProtocolVersion,
     ParseError, // the parsing of a PDU resulted in an error
-    InsufficientHeaderLength(usize), // the input was too small to contain a valid DIS header; (usize found)
-    InsufficientPduLength(usize, usize), // the input was too small to contain a valid DIS Pdu based on the header and parsing; (usize expected, usize found)
+    InsufficientHeaderLength(u16), // the input was too small to contain a valid DIS header; (u16 found)
+    InsufficientPduLength(u16, u16), // the input was too small to contain a valid DIS Pdu based on the header and parsing; (u16 expected, u16 found)
 }
